+++ resources/js/Components/forms/AddUserForm.tsx
@@ -3,10 +3,8 @@ import axios from "axios";
 import { useState } from "react";
 import { useForm, SubmitHandler } from "react-hook-form";
 import { ToastState } from "../Toast";
-import { CloseX } from "../inputs/CloseX";
-import { TextInput } from "../inputs/TextArea";
-import { DropdownInput } from "../inputs/DropdownInput";
-import { SubmitButton } from "../inputs/SubmitButton";
+import { TextInput, DropdownInput, CloseX, SubmitButton } from "../inputs";
+
 type Inputs = {
     name_first: string;
     name_last: string;
