+++ resources/js/Components/inputs/index.js
